+++ packaging/createmsi.py
@@ -39,49 +39,6 @@ def gen_guid():
     '''
     return str(uuid.uuid4()).upper()
 
-def get_all_modules_from_dir(dirname):
-    '''
-    Get all modules required for Meson build MSI package
-    from directories.
-    '''
-    modname = os.path.basename(dirname)
-    modules = [os.path.splitext(os.path.split(x)[1])[0] for x in glob(os.path.join(dirname, '*'))]
-    modules = ['mesonbuild.' + modname + '.' + x for x in modules if not x.startswith('_')]
-    return modules
-
-def get_more_modules():
-    '''
-        Getter for missing Modules.
-        Python packagers want to be minimal and only copy the things
-        that they can see that being used. They are blind to many things.
-    '''
-    return ['distutils.archive_util',
-            'distutils.cmd',
-            'distutils.config',
-            'distutils.core',
-            'distutils.debug',
-            'distutils.dep_util',
-            'distutils.dir_util',
-            'distutils.dist',
-            'distutils.errors',
-            'distutils.extension',
-            'distutils.fancy_getopt',
-            'distutils.file_util',
-            'distutils.spawn',
-            'distutils.util',
-            'distutils.version',
-            'distutils.command.build_ext',
-            'distutils.command.build',
-            'distutils.command.install',
-            'filecmp',
-            ]
-
-def get_modules():
-    modules = get_all_modules_from_dir('mesonbuild/modules')
-    modules += get_all_modules_from_dir('mesonbuild/scripts')
-    modules += get_more_modules()
-    return modules
-
 class Node:
     '''
        Node to hold path and directory values
@@ -166,7 +123,6 @@ class PackageGenerator:
             if os.path.exists(sdir):
                 shutil.rmtree(sdir)
         main_stage, ninja_stage = self.staging_dirs
-        modules = get_modules()
 
         pyinstaller = shutil.which('pyinstaller')
         if not pyinstaller:
@@ -178,10 +134,9 @@ class PackageGenerator:
             shutil.rmtree(pyinstaller_tmpdir)
         pyinst_cmd = [pyinstaller,
                       '--clean',
+                      '--additional-hooks-dir=packaging',
                       '--distpath',
                       pyinstaller_tmpdir]
-        for m in modules:
-            pyinst_cmd += ['--hidden-import', m]
         pyinst_cmd += ['meson.py']
         subprocess.check_call(pyinst_cmd)
         shutil.move(pyinstaller_tmpdir + '/meson', main_stage)
